Renamer pakker og flytter litt på ting
