--- lib/src/main/kotlin/no/kommune/oslo/automatiserteprosesser/maskinporten/MaskinportenKlientException.kt
@@ -1,3 +1,0 @@
-package no.kommune.oslo.automatiserteprosesser.maskinporten
-
-class MaskinportenKlientException(message: String) : Exception(message)
