--- lib/src/main/kotlin/no/kommune/oslo/automatiserteprosesser/maskinporten/MaskinportenTokenWrapper.kt
@@ -1,6 +1,0 @@
-package no.kommune.oslo.automatiserteprosesser.maskinporten
-
-data class MaskinportenTokenWrapper(val access_token: String,
-                                    val token_type: String,
-                                    val expires_in: String,
-                                    val scope: String )
