+++ lib/src/main/kotlin/no/kommune/oslo/origo/maskinporten/MaskinportenFileKeystoreKonfigurasjon.kt
@@ -1,4 +1,4 @@
-package no.kommune.oslo.automatiserteprosesser.maskinporten
+package no.kommune.oslo.origo.maskinporten
 
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
