+++ lib/src/main/kotlin/no/kommune/oslo/origo/maskinporten/MaskinportenJWTGenerator.kt
@@ -1,4 +1,4 @@
-package no.kommune.oslo.automatiserteprosesser.maskinporten
+package no.kommune.oslo.origo.maskinporten
 
 import com.nimbusds.jose.JWSAlgorithm
 import com.nimbusds.jose.JWSHeader
