+++ lib/src/test/kotlin/no/kommune/oslo/origo/maskinporten/MaskinportenTestUtils.kt
@@ -1,4 +1,4 @@
-package no.kommune.oslo.automatiserteprosesser.maskinporten
+package no.kommune.oslo.origo.maskinporten
 
 class MaskinportenTestUtils {
